Refactor API calls into a utils file. Refactor API key into an untracked file.
