--- lib/components/Main/test-data.js
@@ -1,1952 +1,0 @@
-const colorado =
-  {
-    'response': {
-      'version': '0.1',
-      'termsofService': 'http://www.wunderground.com/weather/api/d/terms.html',
-      'features': {
-        'hourly': 1
-  ,
-        'forecast10day': 1
-  ,
-        'conditions': 1,
-      },
-	                                    }
-  ,	'current_observation': {
-		                                    'image': {
-		                  'url': 'http://icons.wxug.com/graphics/wu2/logo_130x80.png',
-		                  'title': 'Weather Underground',
-		                  'link': 'http://www.wunderground.com',
-		},
-		                                    'display_location': {
-		                  'full': 'Littleton, CO',
-		                  'city': 'Littleton',
-		                  'state': 'CO',
-		                  'state_name': 'Colorado',
-		                  'country': 'US',
-		                  'country_iso3166': 'US',
-		                  'zip': '80122',
-		                  'magic': '1',
-		                  'wmo': '99999',
-		                  'latitude': '39.58000183',
-		                  'longitude': '-104.94999695',
-		                  'elevation': '1707.2',
-		},
-		                                    'observation_location': {
-		                  'full': 'Bella Vista Villas, Centennial, Colorado',
-		                  'city': 'Bella Vista Villas, Centennial',
-		                  'state': 'Colorado',
-		                  'country': 'US',
-		                  'country_iso3166': 'US',
-		                  'latitude': '39.580818',
-		                  'longitude': '-104.939926',
-		                  'elevation': '5679 ft',
-		},
-		                                    'estimated': {
-		},
-		                                    'station_id': 'KCOCENTE35',
-		                                    'observation_time': 'Last Updated on May 31, 2:03 PM MDT',
-		                                    'observation_time_rfc822': 'Wed, 31 May 2017 14:03:43 -0600',
-		                                    'observation_epoch': '1496261023',
-		                                    'local_time_rfc822': 'Wed, 31 May 2017 14:03:59 -0600',
-		                                    'local_epoch': '1496261039',
-		                                    'local_tz_short': 'MDT',
-		                                    'local_tz_long': 'America/Denver',
-		                                    'local_tz_offset': '-0600',
-		                                    'weather': 'Mostly Cloudy',
-		                                    'temperature_string': '78.3 F (25.7 C)',
-		                                    'temp_f': 78.3,
-		                                    'temp_c': 25.7,
-		                                    'relative_humidity': '22%',
-		                                    'wind_string': 'From the SE at 4.3 MPH Gusting to 6.0 MPH',
-		                                    'wind_dir': 'SE',
-		                                    'wind_degrees': 124,
-		                                    'wind_mph': 4.3,
-		                                    'wind_gust_mph': '6.0',
-		                                    'wind_kph': 6.9,
-		                                    'wind_gust_kph': '9.7',
-		                                    'pressure_mb': '1012',
-		                                    'pressure_in': '29.90',
-		                                    'pressure_trend': '-',
-		                                    'dewpoint_string': '36 F (2 C)',
-		                                    'dewpoint_f': 36,
-		                                    'dewpoint_c': 2,
-		                                    'heat_index_string': 'NA',
-		                                    'heat_index_f': 'NA',
-		                                    'heat_index_c': 'NA',
-		                                    'windchill_string': 'NA',
-		                                    'windchill_f': 'NA',
-		                                    'windchill_c': 'NA',
-		                                    'feelslike_string': '78.3 F (25.7 C)',
-		                                    'feelslike_f': '78.3',
-		                                    'feelslike_c': '25.7',
-		                                    'visibility_mi': '10.0',
-		                                    'visibility_km': '16.1',
-		                                    'solarradiation': '--',
-		                                    'UV': '7', 'precip_1hr_string': '0.00 in ( 0 mm)',
-		                                    'precip_1hr_in': '0.00',
-		                                    'precip_1hr_metric': ' 0',
-		                                    'precip_today_string': '0.00 in (0 mm)',
-		                                    'precip_today_in': '0.00',
-		                                    'precip_today_metric': '0',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'forecast_url': 'http://www.wunderground.com/US/CO/Littleton.html',
-		                                    'history_url': 'http://www.wunderground.com/weatherstation/WXDailyHistory.asp?ID=KCOCENTE35',
-		                                    'ob_url': 'http://www.wunderground.com/cgi-bin/findweather/getForecast?query=39.580818,-104.939926',
-		                                    'nowcast': '',
-	                    }
-		,
-	                                    'forecast': {
-		                  'txt_forecast': {
-		                  'date': '1:05 PM MDT',
-		                  'forecastday': [
-		                    {
-		                                    'period': 0,
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'title': 'Wednesday',
-		                                    'fcttext': 'A mix of clouds and sun. High 77F. Winds SSW at 10 to 15 mph.',
-		                                    'fcttext_metric': 'Cloudy skies. A stray shower or thunderstorm is possible. High 26C. Winds NE at 15 to 25 km/h.',
-		                                    'pop': '10',
-		                    }
-		,
-		                    {
-		                                    'period': 1,
-		                                    'icon': 'nt_partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'title': 'Wednesday Night',
-		                                    'fcttext': 'Some clouds. Low 52F. Winds SW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Partly to mostly cloudy. A stray shower or thunderstorm is possible. Low 11C. Winds SSW at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    }
-		,
-		                    {
-		                                    'period': 2,
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'title': 'Thursday',
-		                                    'fcttext': 'Partly cloudy. High 79F. Winds SSW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Partly cloudy skies. High 26C. Winds SSW at 10 to 15 km/h.',
-		                                    'pop': '20',
-		                    }
-		,
-		                    {
-		                                    'period': 3,
-		                                    'icon': 'nt_partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'title': 'Thursday Night',
-		                                    'fcttext': 'Partly cloudy. Low 54F. Winds SSW at 10 to 15 mph.',
-		                                    'fcttext_metric': 'Partly cloudy skies. Low 12C. Winds SSW at 15 to 25 km/h.',
-		                                    'pop': '10',
-		                    }
-		,
-		                    {
-		                                    'period': 4,
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'title': 'Friday',
-		                                    'fcttext': 'Sunshine and clouds mixed. A stray shower or thunderstorm is possible. High 79F. Winds NNW at 10 to 15 mph.',
-		                                    'fcttext_metric': 'Partly cloudy skies. A stray shower or thunderstorm is possible. High 26C. Winds N at 10 to 15 km/h.',
-		                                    'pop': '20',
-		                    }
-		,
-		                    {
-		                                    'period': 5,
-		                                    'icon': 'nt_partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'title': 'Friday Night',
-		                                    'fcttext': 'Partly cloudy skies. A stray shower or thunderstorm is possible. Low 54F. Winds SW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Some clouds early will give way to generally clear conditions overnight. A stray shower or thunderstorm is possible. Low 12C. Winds SW at 10 to 15 km/h.',
-		                                    'pop': '20',
-		                    }
-		,
-		                    {
-		                                    'period': 6,
-		                                    'icon': 'clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		                                    'title': 'Saturday',
-		                                    'fcttext': 'Sunny. High 82F. Winds NE at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Mainly sunny. High 28C. Winds NE at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    }
-		,
-		                    {
-		                                    'period': 7,
-		                                    'icon': 'nt_clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_clear.gif',
-		                                    'title': 'Saturday Night',
-		                                    'fcttext': 'Clear. Low near 55F. Winds S at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Clear skies. Low 13C. Winds S at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    }
-		,
-		                    {
-		                                    'period': 8,
-		                                    'icon': 'clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		                                    'title': 'Sunday',
-		                                    'fcttext': 'Except for a few afternoon clouds, mainly sunny. A stray shower or thunderstorm is possible. High 84F. SSW winds shifting to NNE at 10 to 15 mph.',
-		                                    'fcttext_metric': 'Generally sunny despite a few afternoon clouds. A stray shower or thunderstorm is possible. High 29C. Winds NNE at 10 to 15 km/h.',
-		                                    'pop': '20',
-		                    }
-		,
-		                    {
-		                                    'period': 9,
-		                                    'icon': 'nt_clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_clear.gif',
-		                                    'title': 'Sunday Night',
-		                                    'fcttext': 'Clear skies. Low 56F. Winds SW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Clear skies. Low 13C. Winds SW at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    }
-		,
-		                    {
-		                                    'period': 10,
-		                                    'icon': 'chancetstorms',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/chancetstorms.gif',
-		                                    'title': 'Monday',
-		                                    'fcttext': 'Sunny in the morning then increasing clouds with some scattered thunderstorms later in the day. High 82F. Winds NNW at 10 to 15 mph. Chance of rain 50%.',
-		                                    'fcttext_metric': 'Sunny skies during the morning hours. Scattered showers and thunderstorms developing in the afternoon. High 28C. Winds NNW at 10 to 15 km/h. Chance of rain 40%.',
-		                                    'pop': '50',
-		                    }
-		,
-		                    {
-		                                    'period': 11,
-		                                    'icon': 'nt_mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_mostlycloudy.gif',
-		                                    'title': 'Monday Night',
-		                                    'fcttext': 'Partly cloudy skies during the evening will give way to cloudy skies overnight. A stray shower or thunderstorm is possible. Low 56F. Winds SSW at 10 to 15 mph.',
-		                                    'fcttext_metric': 'Considerable cloudiness. A stray shower or thunderstorm is possible. Low 13C. Winds SSW at 10 to 15 km/h.',
-		                                    'pop': '20',
-		                    }
-		,
-		                    {
-		                                    'period': 12,
-		                                    'icon': 'tstorms',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/tstorms.gif',
-		                                    'title': 'Tuesday',
-		                                    'fcttext': 'Scattered thunderstorms in the morning becoming more widespread in the afternoon. High 74F. Winds NNE at 5 to 10 mph. Chance of rain 80%.',
-		                                    'fcttext_metric': 'Scattered thunderstorms in the morning becoming more widespread in the afternoon. High 24C. Winds NNE at 10 to 15 km/h. Chance of rain 80%.',
-		                                    'pop': '80',
-		                    }
-		,
-		                    {
-		                                    'period': 13,
-		                                    'icon': 'nt_tstorms',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_tstorms.gif',
-		                                    'title': 'Tuesday Night',
-		                                    'fcttext': 'Thunderstorms during the evening will give way to steady rain overnight. Low 54F. Winds SSE at 5 to 10 mph. Chance of rain 80%.',
-		                                    'fcttext_metric': 'Thunderstorms in the evening, then cloudy with rain likely overnight. Low 12C. Winds SSE at 10 to 15 km/h. Chance of rain 80%.',
-		                                    'pop': '80',
-		                    }
-		,
-		                    {
-		                                    'period': 14,
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'title': 'Wednesday',
-		                                    'fcttext': 'Sunshine and clouds mixed. High 77F. SW winds shifting to NNE at 10 to 15 mph.',
-		                                    'fcttext_metric': 'Partly cloudy skies. High around 25C. Winds NNE at 10 to 15 km/h.',
-		                                    'pop': '10',
-		                    }
-		,
-		                    {
-		                                    'period': 15,
-		                                    'icon': 'nt_partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'title': 'Wednesday Night',
-		                                    'fcttext': 'Partly cloudy skies. Low 53F. Winds WSW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Partly cloudy. Low 11C. Winds SW at 10 to 15 km/h.',
-		                                    'pop': '10',
-		                    }
-		,
-		                    {
-		                                    'period': 16,
-		                                    'icon': 'clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		                                    'title': 'Thursday',
-		                                    'fcttext': 'Generally sunny despite a few afternoon clouds. High around 80F. Winds NNW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Except for a few afternoon clouds, mainly sunny. High 26C. Winds NNW at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    }
-		,
-		                    {
-		                                    'period': 17,
-		                                    'icon': 'nt_partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'title': 'Thursday Night',
-		                                    'fcttext': 'Some clouds early will give way to generally clear conditions overnight. Low 54F. Winds WSW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Some clouds early will give way to generally clear conditions overnight. Low 12C. Winds WSW at 10 to 15 km/h.',
-		                                    'pop': '10',
-		                    }
-		,
-		                    {
-		                                    'period': 18,
-		                                    'icon': 'clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		                                    'title': 'Friday',
-		                                    'fcttext': 'Except for a few afternoon clouds, mainly sunny. High 84F. Winds NNW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'Except for a few afternoon clouds, mainly sunny. High 29C. Winds NNW at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    }
-		,
-		                    {
-		                                    'period': 19,
-		                                    'icon': 'nt_partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'title': 'Friday Night',
-		                                    'fcttext': 'Partly cloudy skies. Low 54F. Winds SW at 5 to 10 mph.',
-		                                    'fcttext_metric': 'A few clouds from time to time. Low 12C. Winds SW at 10 to 15 km/h.',
-		                                    'pop': '0',
-		                    },
-		],
-		},
-		                  'simpleforecast': {
-		                  'forecastday': [
-		{ 'date': {
-	                  'epoch': '1496278800',
-	                  'pretty': '7:00 PM MDT on May 31, 2017',
-	                  'day': 31,
-	                  'month': 5,
-	                  'year': 2017,
-	                  'yday': 150,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'May',
-	                  'monthname_short': 'May',
-	                  'weekday_short': 'Wed',
-	                  'weekday': 'Wednesday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 1,
-		'high': {
-		                  'fahrenheit': '77',
-		                  'celsius': '25',
-		},
-		'low': {
-		                  'fahrenheit': '52',
-		                  'celsius': '11',
-		},
-		'conditions': 'Partly Cloudy',
-		'icon': 'partlycloudy',
-		'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		'skyicon': '',
-		'pop': 10,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 15,
-		                  'kph': 24,
-		                  'dir': 'SSW',
-		                  'degrees': 213,
-		},
-		'avewind': {
-		                  'mph': 12,
-		                  'kph': 19,
-		                  'dir': 'SSW',
-		                  'degrees': 213,
-		},
-		'avehumidity': 35,
-		'maxhumidity': 56,
-		'minhumidity': 23,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496365200',
-	                  'pretty': '7:00 PM MDT on June 01, 2017',
-	                  'day': 1,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 151,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Thu',
-	                  'weekday': 'Thursday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 2,
-		'high': {
-		                  'fahrenheit': '79',
-		                  'celsius': '26',
-		},
-		'low': {
-		                  'fahrenheit': '54',
-		                  'celsius': '12',
-		},
-		'conditions': 'Partly Cloudy',
-		'icon': 'partlycloudy',
-		'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		'skyicon': '',
-		'pop': 20,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'SSW',
-		                  'degrees': 209,
-		},
-		'avewind': {
-		                  'mph': 9,
-		                  'kph': 14,
-		                  'dir': 'SSW',
-		                  'degrees': 209,
-		},
-		'avehumidity': 46,
-		'maxhumidity': 69,
-		'minhumidity': 23,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496451600',
-	                  'pretty': '7:00 PM MDT on June 02, 2017',
-	                  'day': 2,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 152,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Fri',
-	                  'weekday': 'Friday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 3,
-		'high': {
-		                  'fahrenheit': '79',
-		                  'celsius': '26',
-		},
-		'low': {
-		                  'fahrenheit': '54',
-		                  'celsius': '12',
-		},
-		'conditions': 'Partly Cloudy',
-		'icon': 'partlycloudy',
-		'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		'skyicon': '',
-		'pop': 20,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 15,
-		                  'kph': 24,
-		                  'dir': 'NNW',
-		                  'degrees': 346,
-		},
-		'avewind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNW',
-		                  'degrees': 346,
-		},
-		'avehumidity': 44,
-		'maxhumidity': 66,
-		'minhumidity': 25,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496538000',
-	                  'pretty': '7:00 PM MDT on June 03, 2017',
-	                  'day': 3,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 153,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Sat',
-	                  'weekday': 'Saturday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 4,
-		'high': {
-		                  'fahrenheit': '82',
-		                  'celsius': '28',
-		},
-		'low': {
-		                  'fahrenheit': '55',
-		                  'celsius': '13',
-		},
-		'conditions': 'Clear',
-		'icon': 'clear',
-		'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		'skyicon': '',
-		'pop': 0,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NE',
-		                  'degrees': 44,
-		},
-		'avewind': {
-		                  'mph': 8,
-		                  'kph': 13,
-		                  'dir': 'NE',
-		                  'degrees': 44,
-		},
-		'avehumidity': 39,
-		'maxhumidity': 59,
-		'minhumidity': 21,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496624400',
-	                  'pretty': '7:00 PM MDT on June 04, 2017',
-	                  'day': 4,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 154,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Sun',
-	                  'weekday': 'Sunday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 5,
-		'high': {
-		                  'fahrenheit': '84',
-		                  'celsius': '29',
-		},
-		'low': {
-		                  'fahrenheit': '56',
-		                  'celsius': '13',
-		},
-		'conditions': 'Clear',
-		'icon': 'clear',
-		'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		'skyicon': '',
-		'pop': 20,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 15,
-		                  'kph': 24,
-		                  'dir': 'NNE',
-		                  'degrees': 22,
-		},
-		'avewind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNE',
-		                  'degrees': 22,
-		},
-		'avehumidity': 39,
-		'maxhumidity': 64,
-		'minhumidity': 20,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496710800',
-	                  'pretty': '7:00 PM MDT on June 05, 2017',
-	                  'day': 5,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 155,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Mon',
-	                  'weekday': 'Monday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 6,
-		'high': {
-		                  'fahrenheit': '82',
-		                  'celsius': '28',
-		},
-		'low': {
-		                  'fahrenheit': '56',
-		                  'celsius': '13',
-		},
-		'conditions': 'Chance of a Thunderstorm',
-		'icon': 'chancetstorms',
-		'icon_url': 'http://icons.wxug.com/i/c/k/chancetstorms.gif',
-		'skyicon': '',
-		'pop': 50,
-		'qpf_allday': {
-		                  'in': 0.05,
-		                  'mm': 1,
-		},
-		'qpf_day': {
-		                  'in': 0.05,
-		                  'mm': 1,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 15,
-		                  'kph': 24,
-		                  'dir': 'NNW',
-		                  'degrees': 346,
-		},
-		'avewind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNW',
-		                  'degrees': 346,
-		},
-		'avehumidity': 45,
-		'maxhumidity': 66,
-		'minhumidity': 26,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496797200',
-	                  'pretty': '7:00 PM MDT on June 06, 2017',
-	                  'day': 6,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 156,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Tue',
-	                  'weekday': 'Tuesday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 7,
-		'high': {
-		                  'fahrenheit': '74',
-		                  'celsius': '23',
-		},
-		'low': {
-		                  'fahrenheit': '54',
-		                  'celsius': '12',
-		},
-		'conditions': 'Thunderstorm',
-		'icon': 'tstorms',
-		'icon_url': 'http://icons.wxug.com/i/c/k/tstorms.gif',
-		'skyicon': '',
-		'pop': 80,
-		'qpf_allday': {
-		                  'in': 0.67,
-		                  'mm': 17,
-		},
-		'qpf_day': {
-		                  'in': 0.29,
-		                  'mm': 7,
-		},
-		'qpf_night': {
-		                  'in': 0.38,
-		                  'mm': 10,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNE',
-		                  'degrees': 28,
-		},
-		'avewind': {
-		                  'mph': 8,
-		                  'kph': 13,
-		                  'dir': 'NNE',
-		                  'degrees': 28,
-		},
-		'avehumidity': 61,
-		'maxhumidity': 79,
-		'minhumidity': 44,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496883600',
-	                  'pretty': '7:00 PM MDT on June 07, 2017',
-	                  'day': 7,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 157,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Wed',
-	                  'weekday': 'Wednesday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 8,
-		'high': {
-		                  'fahrenheit': '77',
-		                  'celsius': '25',
-		},
-		'low': {
-		                  'fahrenheit': '53',
-		                  'celsius': '12',
-		},
-		'conditions': 'Partly Cloudy',
-		'icon': 'partlycloudy',
-		'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		'skyicon': '',
-		'pop': 10,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 15,
-		                  'kph': 24,
-		                  'dir': 'NNE',
-		                  'degrees': 16,
-		},
-		'avewind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNE',
-		                  'degrees': 16,
-		},
-		'avehumidity': 58,
-		'maxhumidity': 81,
-		'minhumidity': 35,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1496970000',
-	                  'pretty': '7:00 PM MDT on June 08, 2017',
-	                  'day': 8,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 158,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Thu',
-	                  'weekday': 'Thursday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 9,
-		'high': {
-		                  'fahrenheit': '80',
-		                  'celsius': '27',
-		},
-		'low': {
-		                  'fahrenheit': '54',
-		                  'celsius': '12',
-		},
-		'conditions': 'Clear',
-		'icon': 'clear',
-		'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		'skyicon': '',
-		'pop': 0,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNW',
-		                  'degrees': 346,
-		},
-		'avewind': {
-		                  'mph': 9,
-		                  'kph': 14,
-		                  'dir': 'NNW',
-		                  'degrees': 346,
-		},
-		'avehumidity': 50,
-		'maxhumidity': 74,
-		'minhumidity': 27,
-		}
-		,
-		{ 'date': {
-	                  'epoch': '1497056400',
-	                  'pretty': '7:00 PM MDT on June 09, 2017',
-	                  'day': 9,
-	                  'month': 6,
-	                  'year': 2017,
-	                  'yday': 159,
-	                  'hour': 19,
-	                  'min': '00',
-	                  'sec': 0,
-	                  'isdst': '1',
-	                  'monthname': 'June',
-	                  'monthname_short': 'Jun',
-	                  'weekday_short': 'Fri',
-	                  'weekday': 'Friday',
-	                  'ampm': 'PM',
-	                  'tz_short': 'MDT',
-	                  'tz_long': 'America/Denver',
-},
-		'period': 10,
-		'high': {
-		                  'fahrenheit': '84',
-		                  'celsius': '29',
-		},
-		'low': {
-		                  'fahrenheit': '54',
-		                  'celsius': '12',
-		},
-		'conditions': 'Clear',
-		'icon': 'clear',
-		'icon_url': 'http://icons.wxug.com/i/c/k/clear.gif',
-		'skyicon': '',
-		'pop': 0,
-		'qpf_allday': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_day': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'qpf_night': {
-		                  'in': 0.00,
-		                  'mm': 0,
-		},
-		'snow_allday': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_day': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'snow_night': {
-		                  'in': 0.0,
-		                  'cm': 0.0,
-		},
-		'maxwind': {
-		                  'mph': 10,
-		                  'kph': 16,
-		                  'dir': 'NNW',
-		                  'degrees': 344,
-		},
-		'avewind': {
-		                  'mph': 7,
-		                  'kph': 11,
-		                  'dir': 'NNW',
-		                  'degrees': 344,
-		},
-		'avehumidity': 40,
-		'maxhumidity': 65,
-		'minhumidity': 19,
-		},
-		],
-		},
-	}
-		,
-	                                    'hourly_forecast': [
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '15', 'hour_padded': '15', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496264400', 'pretty': '3:00 PM MDT on May 31, 2017', 'civil': '3:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '76', 'metric': '24' },
-		                                    'dewpoint': { 'english': '36', 'metric': '2' },
-		                                    'condition': 'Overcast',
-		                                    'icon': 'cloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/cloudy.gif',
-		                                    'fctcode': '4',
-		                                    'sky': '90',
-		                                    'wspd': { 'english': '10', 'metric': '16' },
-		                                    'wdir': { 'dir': 'N', 'degrees': '1' },
-		                                    'wx': 'Cloudy',
-		                                    'uvi': '3',
-		                                    'humidity': '23',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '76', 'metric': '24' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '15',
-		                                    'mslp': { 'english': '29.79', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '16', 'hour_padded': '16', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496268000', 'pretty': '4:00 PM MDT on May 31, 2017', 'civil': '4:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '74', 'metric': '23' },
-		                                    'dewpoint': { 'english': '37', 'metric': '3' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '79',
-		                                    'wspd': { 'english': '11', 'metric': '18' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '208' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '3',
-		                                    'humidity': '26',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '74', 'metric': '23' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '15',
-		                                    'mslp': { 'english': '29.78', 'metric': '1008' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '17', 'hour_padded': '17', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496271600', 'pretty': '5:00 PM MDT on May 31, 2017', 'civil': '5:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '74', 'metric': '23' },
-		                                    'dewpoint': { 'english': '38', 'metric': '3' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '77',
-		                                    'wspd': { 'english': '13', 'metric': '21' },
-		                                    'wdir': { 'dir': 'SSE', 'degrees': '159' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '2',
-		                                    'humidity': '26',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '74', 'metric': '23' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '15',
-		                                    'mslp': { 'english': '29.78', 'metric': '1008' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '18', 'hour_padded': '18', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496275200', 'pretty': '6:00 PM MDT on May 31, 2017', 'civil': '6:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '73', 'metric': '23' },
-		                                    'dewpoint': { 'english': '39', 'metric': '4' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '63',
-		                                    'wspd': { 'english': '12', 'metric': '19' },
-		                                    'wdir': { 'dir': 'S', 'degrees': '190' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '1',
-		                                    'humidity': '29',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '73', 'metric': '23' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '15',
-		                                    'mslp': { 'english': '29.79', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '19', 'hour_padded': '19', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496278800', 'pretty': '7:00 PM MDT on May 31, 2017', 'civil': '7:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '72', 'metric': '22' },
-		                                    'dewpoint': { 'english': '40', 'metric': '4' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '73',
-		                                    'wspd': { 'english': '9', 'metric': '14' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '206' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '32',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '72', 'metric': '22' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '15',
-		                                    'mslp': { 'english': '29.81', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '20', 'hour_padded': '20', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496282400', 'pretty': '8:00 PM MDT on May 31, 2017', 'civil': '8:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '68', 'metric': '20' },
-		                                    'dewpoint': { 'english': '42', 'metric': '6' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '79',
-		                                    'wspd': { 'english': '6', 'metric': '10' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '206' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '39',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '68', 'metric': '20' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '15',
-		                                    'mslp': { 'english': '29.82', 'metric': '1010' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '21', 'hour_padded': '21', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496286000', 'pretty': '9:00 PM MDT on May 31, 2017', 'civil': '9:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '65', 'metric': '18' },
-		                                    'dewpoint': { 'english': '43', 'metric': '6' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '62',
-		                                    'wspd': { 'english': '6', 'metric': '10' },
-		                                    'wdir': { 'dir': 'SW', 'degrees': '215' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '45',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '65', 'metric': '18' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '0',
-		                                    'mslp': { 'english': '29.84', 'metric': '1011' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '22', 'hour_padded': '22', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496289600', 'pretty': '10:00 PM MDT on May 31, 2017', 'civil': '10:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '62', 'metric': '17' },
-		                                    'dewpoint': { 'english': '45', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '50',
-		                                    'wspd': { 'english': '7', 'metric': '11' },
-		                                    'wdir': { 'dir': 'NW', 'degrees': '306' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '52',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '62', 'metric': '17' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '0',
-		                                    'mslp': { 'english': '29.86', 'metric': '1011' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '23', 'hour_padded': '23', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '5', 'mon_padded': '05', 'mon_abbrev': 'May', 'mday': '31', 'mday_padded': '31', 'yday': '150', 'isdst': '1', 'epoch': '1496293200', 'pretty': '11:00 PM MDT on May 31, 2017', 'civil': '11:00 PM', 'month_name': 'May', 'month_name_abbrev': 'May', 'weekday_name': 'Wednesday', 'weekday_name_night': 'Wednesday Night', 'weekday_name_abbrev': 'Wed', 'weekday_name_unlang': 'Wednesday', 'weekday_name_night_unlang': 'Wednesday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '60', 'metric': '16' },
-		                                    'dewpoint': { 'english': '44', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '44',
-		                                    'wspd': { 'english': '7', 'metric': '11' },
-		                                    'wdir': { 'dir': 'W', 'degrees': '267' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '56',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '60', 'metric': '16' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '1',
-		                                    'mslp': { 'english': '29.87', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '0', 'hour_padded': '00', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496296800', 'pretty': '12:00 AM MDT on June 01, 2017', 'civil': '12:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '59', 'metric': '15' },
-		                                    'dewpoint': { 'english': '45', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '33',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'SW', 'degrees': '231' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '60',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '59', 'metric': '15' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '1',
-		                                    'mslp': { 'english': '29.88', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '1', 'hour_padded': '01', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496300400', 'pretty': '1:00 AM MDT on June 01, 2017', 'civil': '1:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '57', 'metric': '14' },
-		                                    'dewpoint': { 'english': '45', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '31',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '208' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '63',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '57', 'metric': '14' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.89', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '2', 'hour_padded': '02', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496304000', 'pretty': '2:00 AM MDT on June 01, 2017', 'civil': '2:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '56', 'metric': '13' },
-		                                    'dewpoint': { 'english': '45', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '36',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '208' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '65',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '56', 'metric': '13' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.89', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '3', 'hour_padded': '03', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496307600', 'pretty': '3:00 AM MDT on June 01, 2017', 'civil': '3:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '55', 'metric': '13' },
-		                                    'dewpoint': { 'english': '44', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '37',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '195' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '67',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '55', 'metric': '13' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.88', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '4', 'hour_padded': '04', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496311200', 'pretty': '4:00 AM MDT on June 01, 2017', 'civil': '4:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '54', 'metric': '12' },
-		                                    'dewpoint': { 'english': '44', 'metric': '7' },
-		                                    'condition': 'Clear',
-		                                    'icon': 'clear',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_clear.gif',
-		                                    'fctcode': '1',
-		                                    'sky': '27',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'S', 'degrees': '190' },
-		                                    'wx': 'Mostly Clear',
-		                                    'uvi': '0',
-		                                    'humidity': '68',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '54', 'metric': '12' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.87', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '5', 'hour_padded': '05', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496314800', 'pretty': '5:00 AM MDT on June 01, 2017', 'civil': '5:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '53', 'metric': '12' },
-		                                    'dewpoint': { 'english': '43', 'metric': '6' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '35',
-		                                    'wspd': { 'english': '4', 'metric': '6' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '193' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '69',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '53', 'metric': '12' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.87', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '6', 'hour_padded': '06', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496318400', 'pretty': '6:00 AM MDT on June 01, 2017', 'civil': '6:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '53', 'metric': '12' },
-		                                    'dewpoint': { 'english': '43', 'metric': '6' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '44',
-		                                    'wspd': { 'english': '4', 'metric': '6' },
-		                                    'wdir': { 'dir': 'S', 'degrees': '177' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '67',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '53', 'metric': '12' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '1',
-		                                    'mslp': { 'english': '29.88', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '7', 'hour_padded': '07', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496322000', 'pretty': '7:00 AM MDT on June 01, 2017', 'civil': '7:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '56', 'metric': '13' },
-		                                    'dewpoint': { 'english': '44', 'metric': '7' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '52',
-		                                    'wspd': { 'english': '4', 'metric': '6' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '200' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '63',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '56', 'metric': '13' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '1',
-		                                    'mslp': { 'english': '29.88', 'metric': '1012' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '8', 'hour_padded': '08', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496325600', 'pretty': '8:00 AM MDT on June 01, 2017', 'civil': '8:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '62', 'metric': '17' },
-		                                    'dewpoint': { 'english': '43', 'metric': '6' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '48',
-		                                    'wspd': { 'english': '3', 'metric': '5' },
-		                                    'wdir': { 'dir': 'SW', 'degrees': '226' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '1',
-		                                    'humidity': '51',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '62', 'metric': '17' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '0',
-		                                    'mslp': { 'english': '29.86', 'metric': '1011' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '9', 'hour_padded': '09', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496329200', 'pretty': '9:00 AM MDT on June 01, 2017', 'civil': '9:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '67', 'metric': '19' },
-		                                    'dewpoint': { 'english': '42', 'metric': '6' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '51',
-		                                    'wspd': { 'english': '3', 'metric': '5' },
-		                                    'wdir': { 'dir': 'SW', 'degrees': '229' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '3',
-		                                    'humidity': '40',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '67', 'metric': '19' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '0',
-		                                    'mslp': { 'english': '29.85', 'metric': '1011' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '10', 'hour_padded': '10', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496332800', 'pretty': '10:00 AM MDT on June 01, 2017', 'civil': '10:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '72', 'metric': '22' },
-		                                    'dewpoint': { 'english': '41', 'metric': '5' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '48',
-		                                    'wspd': { 'english': '4', 'metric': '6' },
-		                                    'wdir': { 'dir': 'SW', 'degrees': '227' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '5',
-		                                    'humidity': '33',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '72', 'metric': '22' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '0',
-		                                    'mslp': { 'english': '29.83', 'metric': '1010' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '11', 'hour_padded': '11', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496336400', 'pretty': '11:00 AM MDT on June 01, 2017', 'civil': '11:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '75', 'metric': '24' },
-		                                    'dewpoint': { 'english': '40', 'metric': '4' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '56',
-		                                    'wspd': { 'english': '6', 'metric': '10' },
-		                                    'wdir': { 'dir': 'ENE', 'degrees': '70' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '6',
-		                                    'humidity': '28',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '75', 'metric': '24' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '0',
-		                                    'mslp': { 'english': '29.8', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '12', 'hour_padded': '12', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496340000', 'pretty': '12:00 PM MDT on June 01, 2017', 'civil': '12:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '77', 'metric': '25' },
-		                                    'dewpoint': { 'english': '38', 'metric': '3' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '57',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'ENE', 'degrees': '66' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '8',
-		                                    'humidity': '24',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '77', 'metric': '25' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '6',
-		                                    'mslp': { 'english': '29.78', 'metric': '1008' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '13', 'hour_padded': '13', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496343600', 'pretty': '1:00 PM MDT on June 01, 2017', 'civil': '1:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '78', 'metric': '26' },
-		                                    'dewpoint': { 'english': '37', 'metric': '3' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '56',
-		                                    'wspd': { 'english': '6', 'metric': '10' },
-		                                    'wdir': { 'dir': 'ENE', 'degrees': '77' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '8',
-		                                    'humidity': '23',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '78', 'metric': '26' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '6',
-		                                    'mslp': { 'english': '29.75', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '14', 'hour_padded': '14', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496347200', 'pretty': '2:00 PM MDT on June 01, 2017', 'civil': '2:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '78', 'metric': '26' },
-		                                    'dewpoint': { 'english': '39', 'metric': '4' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '60',
-		                                    'wspd': { 'english': '8', 'metric': '13' },
-		                                    'wdir': { 'dir': 'NNE', 'degrees': '16' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '8',
-		                                    'humidity': '25',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '78', 'metric': '26' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '9',
-		                                    'mslp': { 'english': '29.74', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '15', 'hour_padded': '15', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496350800', 'pretty': '3:00 PM MDT on June 01, 2017', 'civil': '3:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '76', 'metric': '24' },
-		                                    'dewpoint': { 'english': '38', 'metric': '3' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '52',
-		                                    'wspd': { 'english': '9', 'metric': '14' },
-		                                    'wdir': { 'dir': 'S', 'degrees': '174' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '6',
-		                                    'humidity': '25',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '76', 'metric': '24' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '16',
-		                                    'mslp': { 'english': '29.73', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '16', 'hour_padded': '16', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496354400', 'pretty': '4:00 PM MDT on June 01, 2017', 'civil': '4:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '73', 'metric': '23' },
-		                                    'dewpoint': { 'english': '41', 'metric': '5' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '69',
-		                                    'wspd': { 'english': '10', 'metric': '16' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '211' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '4',
-		                                    'humidity': '32',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '73', 'metric': '23' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '20',
-		                                    'mslp': { 'english': '29.74', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '17', 'hour_padded': '17', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496358000', 'pretty': '5:00 PM MDT on June 01, 2017', 'civil': '5:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '73', 'metric': '23' },
-		                                    'dewpoint': { 'english': '41', 'metric': '5' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '64',
-		                                    'wspd': { 'english': '9', 'metric': '14' },
-		                                    'wdir': { 'dir': 'WSW', 'degrees': '243' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '2',
-		                                    'humidity': '32',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '73', 'metric': '23' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '3',
-		                                    'mslp': { 'english': '29.73', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '18', 'hour_padded': '18', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496361600', 'pretty': '6:00 PM MDT on June 01, 2017', 'civil': '6:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '71', 'metric': '22' },
-		                                    'dewpoint': { 'english': '42', 'metric': '6' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '72',
-		                                    'wspd': { 'english': '10', 'metric': '16' },
-		                                    'wdir': { 'dir': 'W', 'degrees': '272' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '1',
-		                                    'humidity': '35',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '71', 'metric': '22' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '4',
-		                                    'mslp': { 'english': '29.74', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '19', 'hour_padded': '19', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496365200', 'pretty': '7:00 PM MDT on June 01, 2017', 'civil': '7:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '70', 'metric': '21' },
-		                                    'dewpoint': { 'english': '43', 'metric': '6' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '71',
-		                                    'wspd': { 'english': '11', 'metric': '18' },
-		                                    'wdir': { 'dir': 'W', 'degrees': '271' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '37',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '70', 'metric': '21' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.74', 'metric': '1007' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '20', 'hour_padded': '20', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496368800', 'pretty': '8:00 PM MDT on June 01, 2017', 'civil': '8:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '68', 'metric': '20' },
-		                                    'dewpoint': { 'english': '45', 'metric': '7' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '67',
-		                                    'wspd': { 'english': '11', 'metric': '18' },
-		                                    'wdir': { 'dir': 'WSW', 'degrees': '252' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '43',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '68', 'metric': '20' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '9',
-		                                    'mslp': { 'english': '29.76', 'metric': '1008' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '21', 'hour_padded': '21', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496372400', 'pretty': '9:00 PM MDT on June 01, 2017', 'civil': '9:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '65', 'metric': '18' },
-		                                    'dewpoint': { 'english': '46', 'metric': '8' },
-		                                    'condition': 'Mostly Cloudy',
-		                                    'icon': 'mostlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_mostlycloudy.gif',
-		                                    'fctcode': '3',
-		                                    'sky': '70',
-		                                    'wspd': { 'english': '9', 'metric': '14' },
-		                                    'wdir': { 'dir': 'WSW', 'degrees': '243' },
-		                                    'wx': 'Mostly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '51',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '65', 'metric': '18' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '5',
-		                                    'mslp': { 'english': '29.78', 'metric': '1008' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '22', 'hour_padded': '22', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496376000', 'pretty': '10:00 PM MDT on June 01, 2017', 'civil': '10:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '63', 'metric': '17' },
-		                                    'dewpoint': { 'english': '46', 'metric': '8' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '57',
-		                                    'wspd': { 'english': '9', 'metric': '14' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '212' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '55',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '63', 'metric': '17' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '9',
-		                                    'mslp': { 'english': '29.8', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '23', 'hour_padded': '23', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '1', 'mday_padded': '01', 'yday': '151', 'isdst': '1', 'epoch': '1496379600', 'pretty': '11:00 PM MDT on June 01, 2017', 'civil': '11:00 PM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Thursday', 'weekday_name_night': 'Thursday Night', 'weekday_name_abbrev': 'Thu', 'weekday_name_unlang': 'Thursday', 'weekday_name_night_unlang': 'Thursday Night', 'ampm': 'PM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '61', 'metric': '16' },
-		                                    'dewpoint': { 'english': '46', 'metric': '8' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '49',
-		                                    'wspd': { 'english': '8', 'metric': '13' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '199' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '57',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '61', 'metric': '16' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '1',
-		                                    'mslp': { 'english': '29.8', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '0', 'hour_padded': '00', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '2', 'mday_padded': '02', 'yday': '152', 'isdst': '1', 'epoch': '1496383200', 'pretty': '12:00 AM MDT on June 02, 2017', 'civil': '12:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Friday', 'weekday_name_night': 'Friday Night', 'weekday_name_abbrev': 'Fri', 'weekday_name_unlang': 'Friday', 'weekday_name_night_unlang': 'Friday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '60', 'metric': '16' },
-		                                    'dewpoint': { 'english': '46', 'metric': '8' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '44',
-		                                    'wspd': { 'english': '7', 'metric': '11' },
-		                                    'wdir': { 'dir': 'S', 'degrees': '191' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '59',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '60', 'metric': '16' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '2',
-		                                    'mslp': { 'english': '29.81', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '1', 'hour_padded': '01', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '2', 'mday_padded': '02', 'yday': '152', 'isdst': '1', 'epoch': '1496386800', 'pretty': '1:00 AM MDT on June 02, 2017', 'civil': '1:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Friday', 'weekday_name_night': 'Friday Night', 'weekday_name_abbrev': 'Fri', 'weekday_name_unlang': 'Friday', 'weekday_name_night_unlang': 'Friday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '59', 'metric': '15' },
-		                                    'dewpoint': { 'english': '47', 'metric': '8' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '47',
-		                                    'wspd': { 'english': '6', 'metric': '10' },
-		                                    'wdir': { 'dir': 'S', 'degrees': '191' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '64',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '59', 'metric': '15' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '5',
-		                                    'mslp': { 'english': '29.81', 'metric': '1009' },
-		                    }
-		,
-		                    {
-		                                    'FCTTIME': {
-		                  'hour': '2', 'hour_padded': '02', 'min': '00', 'min_unpadded': '0', 'sec': '0', 'year': '2017', 'mon': '6', 'mon_padded': '06', 'mon_abbrev': 'Jun', 'mday': '2', 'mday_padded': '02', 'yday': '152', 'isdst': '1', 'epoch': '1496390400', 'pretty': '2:00 AM MDT on June 02, 2017', 'civil': '2:00 AM', 'month_name': 'June', 'month_name_abbrev': 'Jun', 'weekday_name': 'Friday', 'weekday_name_night': 'Friday Night', 'weekday_name_abbrev': 'Fri', 'weekday_name_unlang': 'Friday', 'weekday_name_night_unlang': 'Friday Night', 'ampm': 'AM', 'tz': '', 'age': '', 'UTCDATE': '',
-		},
-		                                    'temp': { 'english': '58', 'metric': '14' },
-		                                    'dewpoint': { 'english': '47', 'metric': '8' },
-		                                    'condition': 'Partly Cloudy',
-		                                    'icon': 'partlycloudy',
-		                                    'icon_url': 'http://icons.wxug.com/i/c/k/nt_partlycloudy.gif',
-		                                    'fctcode': '2',
-		                                    'sky': '44',
-		                                    'wspd': { 'english': '5', 'metric': '8' },
-		                                    'wdir': { 'dir': 'SSW', 'degrees': '197' },
-		                                    'wx': 'Partly Cloudy',
-		                                    'uvi': '0',
-		                                    'humidity': '65',
-		                                    'windchill': { 'english': '-9999', 'metric': '-9999' },
-		                                    'heatindex': { 'english': '-9999', 'metric': '-9999' },
-		                                    'feelslike': { 'english': '58', 'metric': '14' },
-		                                    'qpf': { 'english': '0.0', 'metric': '0' },
-		                                    'snow': { 'english': '0.0', 'metric': '0' },
-		                                    'pop': '6',
-		                                    'mslp': { 'english': '29.81', 'metric': '1009' },
-		                    },
-	],
-  };
-
-export default colorado;
