+++ lib/utils/dataUtils.js
@@ -1,4 +1,4 @@
-const scrubData = (data) => {
+const getScrubbedData = data => {
   const hourlyData =
     data.hourly_forecast.map(hour => {
       return { unit: hour.FCTTIME.civil,
@@ -24,15 +24,6 @@ const scrubData = (data) => {
     todayCondition: data.forecast.txt_forecast.forecastday[0].fcttext,
   };
 
-  // this.setState(
-  //   { now: nowData,
-  //     hourly: hourlyData,
-  //     daily: dailyData,
-  //     showHourly: true,
-  //     showDaily: false,
-  //   }
-  // );
-
   return {
     now: nowData,
     hourly: hourlyData,
@@ -42,4 +33,4 @@ const scrubData = (data) => {
   };
 };
 
-export default scrubData;
+export default getScrubbedData;
